+++ lib/models/money_objects/categories/categories.dart
@@ -104,7 +104,7 @@ class Categories extends MoneyObjects<Category> {
         type: type,
       );
 
-      addEntry(category);
+      addEntry(moneyObject: category);
     }
     // TODO
     // else if (result.IsDeleted) {
@@ -212,16 +212,29 @@ class Categories extends MoneyObjects<Category> {
   @override
   void loadDemoData() {
     clear();
-    addEntry(Category(id: 0, name: 'Paychecks', description: '', type: CategoryType.income, color: '#FFAAFFBB'));
-    addEntry(Category(id: 1, name: 'Investment', description: '', type: CategoryType.investment, color: '#FFA1A2A3'));
-    addEntry(Category(id: 2, name: 'Interests', description: '', type: CategoryType.income, color: '#FFFF2233'));
-    addEntry(Category(id: 3, name: 'Rental', description: '', type: CategoryType.income, color: '#FF11FF33'));
-    addEntry(Category(id: 4, name: 'Lottery', description: '', type: CategoryType.none, color: '#FF1122FF'));
-    addEntry(Category(id: 5, name: 'Mortgage', description: '', type: CategoryType.expense, color: '#FFBB2233'));
-    addEntry(Category(id: 6, name: 'Saving', description: '', type: CategoryType.income, color: '#FFBB2233'));
-    addEntry(Category(id: 7, name: 'Bills', description: '', type: CategoryType.expense, color: '#FF11DD33'));
-    addEntry(Category(id: 8, name: 'Taxes', description: '', type: CategoryType.expense, color: '#FF1122DD'));
-    addEntry(Category(id: 9, name: 'School', description: '', type: CategoryType.expense));
+    addEntry(
+        moneyObject:
+            Category(id: 0, name: 'Paychecks', description: '', type: CategoryType.income, color: '#FFAAFFBB'));
+    addEntry(
+        moneyObject:
+            Category(id: 1, name: 'Investment', description: '', type: CategoryType.investment, color: '#FFA1A2A3'));
+    addEntry(
+        moneyObject:
+            Category(id: 2, name: 'Interests', description: '', type: CategoryType.income, color: '#FFFF2233'));
+    addEntry(
+        moneyObject: Category(id: 3, name: 'Rental', description: '', type: CategoryType.income, color: '#FF11FF33'));
+    addEntry(
+        moneyObject: Category(id: 4, name: 'Lottery', description: '', type: CategoryType.none, color: '#FF1122FF'));
+    addEntry(
+        moneyObject:
+            Category(id: 5, name: 'Mortgage', description: '', type: CategoryType.expense, color: '#FFBB2233'));
+    addEntry(
+        moneyObject: Category(id: 6, name: 'Saving', description: '', type: CategoryType.income, color: '#FFBB2233'));
+    addEntry(
+        moneyObject: Category(id: 7, name: 'Bills', description: '', type: CategoryType.expense, color: '#FF11DD33'));
+    addEntry(
+        moneyObject: Category(id: 8, name: 'Taxes', description: '', type: CategoryType.expense, color: '#FF1122DD'));
+    addEntry(moneyObject: Category(id: 9, name: 'School', description: '', type: CategoryType.expense));
   }
 
   @override