+++ lib/models/money_objects/investments/investments.dart
@@ -10,7 +10,7 @@ class Investments extends MoneyObjects<Investment> {
   void loadFromJson(final List<MyJson> rows) {
     clear();
     for (final MyJson row in rows) {
-      addEntry(Investment.fromJson(row));
+      addEntry(moneyObject: Investment.fromJson(row));
     }
   }