+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Fix Web version
 
+## Refactor
+
+- [notifyTransactionChange()| MoneyObject.addEntry()] use named parameters
+
 ## [version 1.4.4] 2024-03-18
 
 ## Fixed